Extract manifest modification functions from playlist loaders to manifest module
After parsing an HLS or DASH manifest string via m3u8-parser or
mpd-parser, the respective playlist loader (playlist-loader or
dash-playlist-loader) would modify the manifest objects to add
properties or change their structure so that VHS could use them
as if they were standardized.

The functions for doing this lived in playlist-loader and
dash-playlist-loader, however, with the upcoming possibility of passing
in a vhs-json object as the source (representing a manifest string
already parsed by m3u8-parser/mpd-parser), the functions needed to be
extracted to be reused appropriately, in a way that could act on either
HLS or DASH playlists in the same fashion.

This extracts that functionality to the manifests.js module.
