+++ test/dash-playlist-loader.test.js
@@ -6,7 +6,8 @@ import {
   requestSidx_,
   generateSidxKey,
   compareSidxEntry,
-  filterChangedSidxMappings
+  filterChangedSidxMappings,
+  parseMasterXml
 } from '../src/dash-playlist-loader';
 import xhrFactory from '../src/xhr';
 import {
@@ -1206,7 +1207,12 @@ QUnit.test('parseMasterXml: setup phony playlists and resolves uris', function(a
   loader.load();
   this.standardXHRResponse(this.requests.shift());
 
-  const masterPlaylist = loader.parseMasterXml();
+  const masterPlaylist = parseMasterXml({
+    masterXml: loader.masterXml_,
+    srcUrl: loader.srcUrl,
+    clientOffset: loader.clientOffset_,
+    sidxMapping: loader.sidxMapping_
+  });
 
   assert.strictEqual(masterPlaylist.uri, loader.srcUrl, 'master playlist uri set correctly');
   assert.strictEqual(masterPlaylist.playlists[0].uri, 'placeholder-uri-0');
@@ -1228,11 +1234,24 @@ QUnit.test('parseMasterXml: includes sidx info if available and matches playlist
 
   loader.load();
   this.standardXHRResponse(this.requests.shift());
-  const origParsedMaster = loader.parseMasterXml();
+  const origParsedMaster = parseMasterXml({
+    masterXml: loader.masterXml_,
+    srcUrl: loader.srcUrl,
+    clientOffset: loader.clientOffset_,
+    sidxMapping: loader.sidxMapping_
+  });
 
   loader.sidxMapping_ = {};
+
+  let newParsedMaster = parseMasterXml({
+    masterXml: loader.masterXml_,
+    srcUrl: loader.srcUrl,
+    clientOffset: loader.clientOffset_,
+    sidxMapping: loader.sidxMapping_
+  });
+
   assert.deepEqual(
-    loader.parseMasterXml(),
+    newParsedMaster,
     origParsedMaster,
     'empty sidxMapping will not affect master xml parsing'
   );
@@ -1253,10 +1272,15 @@ QUnit.test('parseMasterXml: includes sidx info if available and matches playlist
       }]
     }
   };
-  const newMaster = loader.parseMasterXml();
+  newParsedMaster = parseMasterXml({
+    masterXml: loader.masterXml_,
+    srcUrl: loader.srcUrl,
+    clientOffset: loader.clientOffset_,
+    sidxMapping: loader.sidxMapping_
+  });
 
   assert.deepEqual(
-    newMaster.playlists[0].segments[0].byterange,
+    newParsedMaster.playlists[0].segments[0].byterange,
     {
       length: 10,
       offset: 400
@@ -1264,7 +1288,7 @@ QUnit.test('parseMasterXml: includes sidx info if available and matches playlist
     'byte range from sidx is applied to playlist segment'
   );
   assert.deepEqual(
-    newMaster.playlists[0].segments[0].map.byterange,
+    newParsedMaster.playlists[0].segments[0].map.byterange,
     {
       length: 200,
       offset: 0
@@ -1430,7 +1454,12 @@ QUnit.test('refreshXml_: requests the sidx if it changed', function(assert) {
   // child playlist
   this.standardXHRResponse(this.requests.shift());
 
-  const oldMaster = loader.parseMasterXml();
+  const oldMaster = parseMasterXml({
+    masterXml: loader.masterXml_,
+    srcUrl: loader.srcUrl,
+    clientOffset: loader.clientOffset_,
+    sidxMapping: loader.sidxMapping_
+  });
   const newMasterXml = loader.masterXml_.replace(/(indexRange)=\"\d+-\d+\"/g, '$1="400-599"');
 
   loader.masterXml_ = newMasterXml;
@@ -1441,16 +1470,29 @@ QUnit.test('refreshXml_: requests the sidx if it changed', function(assert) {
     },
     'sidx is the original in the xml'
   );
+  let newMaster = parseMasterXml({
+    masterXml: loader.masterXml_,
+    srcUrl: loader.srcUrl,
+    clientOffset: loader.clientOffset_,
+    sidxMapping: loader.sidxMapping_
+  });
+
   assert.notEqual(
-    loader.parseMasterXml().playlists[0].sidx.byterange.offset,
+    newMaster.playlists[0].sidx.byterange.offset,
     oldMaster.playlists[0].sidx.byterange.offset,
     'the sidx has been changed'
   );
   loader.refreshXml_();
 
   assert.strictEqual(this.requests.length, 1, 'manifest is being requested');
+  newMaster = parseMasterXml({
+    masterXml: loader.masterXml_,
+    srcUrl: loader.srcUrl,
+    clientOffset: loader.clientOffset_,
+    sidxMapping: loader.sidxMapping_
+  });
   assert.deepEqual(
-    loader.parseMasterXml().playlists[0].sidx.byterange,
+    newMaster.playlists[0].sidx.byterange,
     {
       offset: 400,
       length: 200
